+++ coordinator/terraform/gcp/modules/keydb/variables.tf
@@ -50,3 +50,13 @@ variable "key_db_retention_period" {
   type        = string
   nullable    = false
 }
+
+variable "backups" {
+  description = "Spanner database backup settings."
+  type = object({
+    # In seconds, e.g. "86400s" for 1 day, max 366 days.
+    retention_duration = string
+    cron_spec          = string
+    incremental        = bool
+  })
+}